perf pmu: When using default config, record which bits of config were changed by the user
Default config is static so cannot be changed, for example, if the user
chooses AUX area sampling mode. In the case of Intel PT, the psb_period
needs to be small for sampling, so there is a need to set the default
psb_period to 0 (2 KiB) in that case. However that should not override a
value set by the user. To allow for that, when using default config, record
which bits of config were changed by the user.

Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>