perf: Disable IRQs in address filter sync path
If PMU callbacks are executed in hardirq context, the address filter
sync code needs to disable interrupts when taking its spinlock, to be
consistent with the rest of its users. This may happen if the PMU is
used in AUX sampling or the PMI is delivered as a regular interrupt.

Signed-off-by: Alexander Shishkin <alexander.shishkin@linux.intel.com>